Push notifications via expirationCheck.js now working without server-side axios calls
