+++ server.js
@@ -4,7 +4,7 @@ const express = require("express");
 const mongoose = require("mongoose");
 const CronJob = require('cron').CronJob;
 const dailyCheck = require('./cronJobs/expirationCheck');
-const routes = require("./routes/routes");
+const routes = require("./routes/routes").router;
 
 const PORT = process.env.PORT || 3001;
 const app = express();
